symex: klee: Fix KLEE build by changing the build type
Otherwise, this fails with an internal error in Clang. The error
occurs while attempting to tools/ktest-gen/ktest-gen.cpp. Noticed
this on the Guix QA. I couldn't reproduce this previously but can
reproduce it now, though only under certain circumstances. For example,
when debugging the build by sourcing the environment-variables file,
the error cannot be reproduced.